Backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.7.1
python-dotenv==1.0.1